+++ cmd/ccd/server/delete.go
@@ -7,6 +7,7 @@ import (
 	"github.com/indexdata/ccms"
 	"github.com/indexdata/ccms/cmd/ccd/ast"
 	"github.com/indexdata/ccms/cmd/ccd/catalog"
+	"github.com/indexdata/ccms/internal/global"
 )
 
 func deleteStmt(s *svr, rqid int64, cmd *ast.DeleteStmt) *ccms.Result {
@@ -24,7 +25,7 @@ func deleteStmt(s *svr, rqid int64, cmd *ast.DeleteStmt) *ccms.Result {
 	}
 	//log.Info("[%d] %s", rqid, sql)
 	if _, err := s.dp.Exec(context.TODO(), sql); err != nil {
-		return cmderr(fmt.Sprintf("deleting from %q: %v", cmd.From, err))
+		return cmderr(fmt.Sprintf("deleting from %q: %v", cmd.From, global.PGErr(err)))
 	}
 
 	return ccms.NewResult("delete")