+++ cmd/ccd/server/info.go
@@ -14,13 +14,14 @@ func infoStmt(s *svr, cmd *ast.InfoStmt) *ccms.Result {
 		result.AddField("info", "text")
 		result.AddData([]any{"" +
 			"SQL commands:\n" +
-			"        create set  define a new set\n" +
-			"        delete      remove objects from set membership\n" +
-			"        drop set    remove a set\n" +
+			"        create set   define a new set\n" +
+			"        create user  define a new user\n" +
+			"        delete       remove objects from set membership\n" +
+			"        drop set     remove a set\n" +
 			//"        info    show supported commands\n" +
-			"        insert      insert objects into a set\n" +
-			"        select      retrieve objects from a set\n" +
-			"        show        list existing filters or sets\n"})
+			"        insert       insert objects into a set\n" +
+			"        select       retrieve objects from a set\n" +
+			"        show         list existing filters or sets\n"})
 		return result
 	}
 
@@ -28,6 +29,8 @@ func infoStmt(s *svr, cmd *ast.InfoStmt) *ccms.Result {
 	switch cmd.Topic {
 	case "create set":
 		docstr = doc.CreateSet()
+	case "create user":
+		docstr = doc.CreateUser()
 	case "delete":
 		docstr = doc.Delete()
 	case "drop set":