+++ internal/global/global.go
@@ -1,6 +1,12 @@
 package global
 
-import "path/filepath"
+import (
+	"errors"
+	"path/filepath"
+	"strings"
+
+	"github.com/jackc/pgx/v5/pgconn"
+)
 
 // Version is defined at build time via -ldflags.
 var Version = ""
@@ -13,4 +19,20 @@ const ClientProgram = "ccc"
 
 func ServerConfigFileName(datadir string) string {
 	return filepath.Join(datadir, "ccd.conf")
+}
+
+func PGErr(err error) error {
+	e := err.(*pgconn.PgError)
+	var b strings.Builder
+	b.WriteString(e.Message)
+	if e.Detail != "" {
+		b.WriteString(": ")
+		b.WriteString(e.Detail)
+	}
+	if e.Hint != "" {
+		b.WriteString(" (")
+		b.WriteString(e.Hint)
+		b.WriteRune(')')
+	}
+	return errors.New(b.String())
 }